Refactor: 알람 api 서비스
