+++ src/main/java/com/openbook/openbook/booth/controller/response/BoothReserveDetailManageResponse.java
@@ -2,7 +2,7 @@ package com.openbook.openbook.booth.controller.response;
 
 import com.openbook.openbook.booth.entity.BoothReservationDetail;
 import com.openbook.openbook.booth.entity.dto.BoothReservationStatus;
-import com.openbook.openbook.user.dto.UserPublicData;
+import com.openbook.openbook.user.controller.response.UserPublicResponse;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 
@@ -11,15 +11,15 @@ public record BoothReserveDetailManageResponse(
         String times,
         @Enumerated(EnumType.STRING)
         BoothReservationStatus status,
-        UserPublicData reserveUser
+        UserPublicResponse reserveUser
 ) {
     public static BoothReserveDetailManageResponse of(BoothReservationDetail detail){
-        UserPublicData userPublicData = detail.getUser() != null ? UserPublicData.of(detail.getUser()) : null;
+        UserPublicResponse userPublicResponse = detail.getUser() != null ? UserPublicResponse.of(detail.getUser()) : null;
         return new BoothReserveDetailManageResponse(
                 detail.getId(),
                 detail.getTime(),
                 detail.getStatus(),
-                userPublicData
+                userPublicResponse
         );
     }
 }
