+++ src/main/java/com/openbook/openbook/booth/service/ManagerBoothService.java
@@ -22,7 +22,7 @@ import com.openbook.openbook.booth.service.core.BoothTagService;
 import com.openbook.openbook.global.exception.ErrorCode;
 import com.openbook.openbook.global.exception.OpenBookException;
 import com.openbook.openbook.user.entity.User;
-import com.openbook.openbook.user.service.core.UserService;
+import com.openbook.openbook.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
