+++ src/main/java/com/openbook/openbook/booth/service/core/BoothReservationDetailService.java
@@ -7,7 +7,6 @@ import com.openbook.openbook.booth.repository.BoothReservationDetailRepository;
 import com.openbook.openbook.global.exception.ErrorCode;
 import com.openbook.openbook.global.exception.OpenBookException;
 import com.openbook.openbook.user.entity.User;
-import com.openbook.openbook.user.service.core.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
