+++ src/main/java/com/openbook/openbook/event/service/AdminEventService.java
@@ -9,8 +9,8 @@ import com.openbook.openbook.event.service.core.EventTagService;
 import com.openbook.openbook.global.exception.ErrorCode;
 import com.openbook.openbook.global.exception.OpenBookException;
 import com.openbook.openbook.user.entity.dto.AlarmType;
-import com.openbook.openbook.user.service.core.AlarmService;
-import com.openbook.openbook.user.service.core.UserService;
+import com.openbook.openbook.user.service.AlarmService;
+import com.openbook.openbook.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
