+++ src/main/java/com/openbook/openbook/event/service/EventCommonService.java
@@ -7,14 +7,11 @@ import com.openbook.openbook.event.dto.EventLayoutCreateData;
 import com.openbook.openbook.booth.dto.BoothAreaCreateData;
 import com.openbook.openbook.event.controller.response.UserEventData;
 import com.openbook.openbook.event.controller.response.EventDetail;
-import com.openbook.openbook.event.controller.response.EventLayoutStatus;
 import com.openbook.openbook.booth.service.core.BoothService;
 import com.openbook.openbook.event.dto.EventDTO;
-import com.openbook.openbook.event.dto.EventNoticeDto;
 import com.openbook.openbook.event.entity.dto.EventStatus;
 import com.openbook.openbook.event.entity.Event;
 import com.openbook.openbook.event.entity.EventLayout;
-import com.openbook.openbook.event.entity.EventTag;
 import com.openbook.openbook.event.service.core.EventNoticeService;
 import com.openbook.openbook.event.service.core.EventService;
 import com.openbook.openbook.event.service.core.EventTagService;
@@ -24,8 +21,8 @@ import com.openbook.openbook.global.exception.OpenBookException;
 import com.openbook.openbook.global.util.TagUtil;
 import com.openbook.openbook.user.entity.dto.AlarmType;
 import com.openbook.openbook.user.entity.User;
-import com.openbook.openbook.user.service.core.AlarmService;
-import com.openbook.openbook.user.service.core.UserService;
+import com.openbook.openbook.user.service.AlarmService;
+import com.openbook.openbook.user.service.UserService;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.stream.IntStream;
