+++ src/main/java/com/openbook/openbook/event/service/ManagerEventService.java
@@ -11,8 +11,7 @@ import com.openbook.openbook.event.service.core.EventService;
 import com.openbook.openbook.event.service.core.EventTagService;
 import com.openbook.openbook.global.exception.ErrorCode;
 import com.openbook.openbook.global.exception.OpenBookException;
-import com.openbook.openbook.global.util.S3Service;
-import com.openbook.openbook.user.service.core.UserService;
+import com.openbook.openbook.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
