+++ src/main/java/com/openbook/openbook/event/service/common/CommonEventReviewService.java
@@ -11,7 +11,7 @@ import com.openbook.openbook.event.service.core.EventService;
 import com.openbook.openbook.global.exception.ErrorCode;
 import com.openbook.openbook.global.exception.OpenBookException;
 import com.openbook.openbook.user.entity.User;
-import com.openbook.openbook.user.service.core.UserService;
+import com.openbook.openbook.user.service.UserService;
 import java.time.LocalDate;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
