+++ src/main/java/com/openbook/openbook/user/controller/response/UserPublicResponse.java
@@ -1,0 +1,17 @@
+package com.openbook.openbook.user.controller.response;
+
+import com.openbook.openbook.user.dto.UserDto;
+
+public record UserPublicResponse(
+        Long id,
+        String nickname,
+        String role
+) {
+    public static UserPublicResponse of(UserDto user) {
+        return new UserPublicResponse(
+                user.id(),
+                user.nickname(),
+                user.role()
+        );
+    }
+}
