--- src/main/java/com/openbook/openbook/user/dto/UserDTO.java
@@ -1,12 +1,0 @@
-package com.openbook.openbook.user.dto;
-
-import lombok.Builder;
-
-@Builder
-public record UserDTO(
-        String email,
-        String password,
-        String name,
-        String nickname
-) {
-}
