+++ src/main/java/com/openbook/openbook/user/dto/UserDto.java
@@ -2,14 +2,16 @@ package com.openbook.openbook.user.dto;
 
 import com.openbook.openbook.user.entity.User;
 
-public record UserPublicData(
+public record UserDto(
         Long id,
+        String name,
         String nickname,
         String role
 ) {
-    public static UserPublicData of(User user) {
-        return new UserPublicData(
+    public static UserDto of(User user) {
+        return new UserDto(
                 user.getId(),
+                user.getName(),
                 user.getNickname(),
                 user.getRole().name()
         );
